Beffermann/requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
appscript==1.1.0
argcomplete==1.10.0
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.3
Beaker==1.6.4
beautifulsoup4==4.8.0
bottle==0.12.7
certifi==2020.4.5.1
cffi==1.13.1
chardet==3.0.4
click==7.1.1
codecov==2.0.22
coverage==4.4.1
dateparser==0.7.4
decorator==4.0.11
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.1.0
django-jet==1.0.8
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
freezegun==0.3.15
gunicorn==20.0.4
httplib2==0.9
idna==2.9
imageio==2.1.2
img2pdf==0.3.3
importlib-metadata==1.6.0
InstagramAPI==1.0.2
jet-bridge-base==0.7.6
jet-django==1.0.1
lxml==4.4.1
mccabe==0.6.1
mock==4.0.2
moviepy==0.2.3.2
multidict==4.7.6
np==1.0.2
numpy==1.17.2
ocrmypdf==9.0.3
packaging==20.3
pandas==0.25.1
pdfminer.six==20181108
pikepdf==1.6.5
Pillow==7.1.2
pluggy==0.13.1
prompt-toolkit==2.0.9
psutil==5.7.0
psycopg2-binary==2.8.5
py==1.8.1
pycodestyle==2.6.0
pycparser==2.19
pycryptodome==3.9.0
pyflakes==2.1.1
PyJWT==1.7.1
pymongo==3.9.0
pyparsing==2.4.7
pytesseract==0.3.4
python-dateutil==2.8.0
python-instagram==1.1.3
pytz==2020.1
redis==2.10.3
regex==2020.5.13
reportlab==3.5.31
requests==2.23.0
requests-toolbelt==0.7.0
selenium==3.141.0
simplejson==3.6.3
six==1.12.0
slackclient==2.7.1
sortedcontainers==2.1.0
soupsieve==2.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
toml==0.10.0
tox==3.15.0
tqdm==4.11.2
tzlocal==2.1
urllib3==1.25.9
virtualenv==20.0.20
wcwidth==0.1.9
whitenoise==5.0.1
XlsxWriter==1.2.1
xlwings==0.19.4
yarl==1.4.2
zipp==3.1.0
